Fix to slicekit flash muxing for v13 tools.
Due to a mapper change in v13, the tile id is no longer 0 when XScope is
enabled. This meant that the flash pin muxing on the slicekit core board was not
being correctly disabled.

This patch now does the correct thing and gets the tile id of
tile 0 and compares it to the current tile id.